Apply PR #9 review suggestions

+++ lexnlp/nlp/en/segments/titles.py
@@ -230,7 +230,7 @@ def get_titles(text, window_pre=3, window_post=3, score_threshold=0.5) -> Genera
 
     # Predict title lines
     # Avoid pandas dtype deprecation noise in sklearn validation by passing a numpy array.
-    predicted_lines = SECTION_SEGMENTER_MODEL.predict_proba(feature_data.to_numpy())
+    predicted_lines = SECTION_SEGMENTER_MODEL.predict_proba(feature_data.to_numpy(dtype=float))
     predicted_df = pandas.DataFrame(predicted_lines, columns=["prob_false", "prob_true"])
     title_lines = predicted_df.loc[predicted_df["prob_true"] >= score_threshold, :].index.tolist()
